Fix: double submission of initial scoring form caused subsequent score entry form to fail
Have fixed a bug where the accidental double submission of the initial scoring form would cause a second set of activity records to be entered for the group, which would cause indexing errors when trying to display the form to enter scores.
Form processing route now checks if any activities have already been entered and silently recovers from this.